% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirem-package.R
\docType{package}
\name{envirem}
\alias{envirem}
\alias{envirem-package}
\title{envirem}
\description{
Generation of bioclimatic rasters that will be particularly useful 
for species distribution modeling.
}
\details{
Package: bioclimExtension\cr
Type: Package\cr
Version: 1.0\cr
Date: 2015-11-01\cr
License: GPL-2 | GPL-3\cr

Temperature rasters are assumed to be in degrees C * 10 and precipitation rasters are 
assumed to be in mm, as it is in Worldclim. 

The main function for generating ENVIREM rasters is \code{\link{generateRasters}}. 
A complete tutorial of this R package can be found at \url{http://envirem.gituhub.io}.
}
\author{
Pascal O. Title, Jordan B. Bemmels
}
\references{
\url{http://envirem.gituhub.io}

Title, P.O. and J.B. Bemmels. \emph{Submitted}. ENVIREM: An expanded set of bioclimatic variables improves ecological niche modeling performance. \emph{Methods in Ecology and Evolution}.
}
\keyword{package}

