% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRasters.R
\name{generateRasters}
\alias{generateRasters}
\title{Execute Layer Creation}
\usage{
generateRasters(var, maindir, resName, timeName, outputDir,
  rasterExt = ".tif", nTiles = 1, overwriteResults = TRUE,
  outputFormat = "GTiff", tempDir = "~/temp", gdalinfoPath = NULL,
  gdal_translatePath = NULL)
}
\arguments{
\item{var}{a vector of variable names to generate, see Details.}

\item{maindir}{path to directory of input rasters}

\item{resName}{output nametag for the resolution}

\item{timeName}{output nametag for the time period}

\item{outputDir}{output directory. A directory will be generated according 
to the resName and timeName, so this is the output location for the 
directory that will be generated.}

\item{rasterExt}{the file extension of the input rasters}

\item{nTiles}{the number of tiles to split the rasters when 
tiling is requested, must be a perfect square}

\item{overwriteResults}{logical, should existing rasters be overwritten}

\item{outputFormat}{output format for rasters, see \code{\link{writeRaster}} for options}

\item{tempDir}{temporary directory that will be created and then removed}

\item{gdalinfoPath}{path to gdalinfo binary, leave as \code{NULL} if it is in the default search path.}

\item{gdal_translatePath}{path to gdal_translate binary, leave as \code{NULL} if it is in the default search path.}
}
\value{
The requested set of rasterLayers will be written to \code{outputDir}.
}
\description{
Main function to generate specified layers for multiple temporal datasets.
If requested, this function will split input rasters into tiles, generate desired variables,
and reassemble the results. Results are named according to specified resName and timeName.
}
\details{
Possible variables to generate include:\cr
aridityIndexThornthwaite \cr
airidityIndexUNEP \cr
climaticMoistureIndex \cr
continentality \cr
embergerQ \cr
growingDegDays0 \cr
growingDegDays5 \cr
humidityIndex \cr
monthCountByTemp10 \cr
PETseasonality \cr
thermicityIndex \cr
minSummerPrecip \cr
maxSummerPrecip \cr
minWinterPrecip \cr
maxWinterPrecip \cr
minTempWarmest \cr
maxTempColdest \cr

If \code{var = 'all'}, then all of the variables will be generated.

\code{resName} and \code{timeName} are only used for naming the output directory.

Rasters in \code{mainDir} should be named appropriately and with identical resolution, origin and extent.
}
\author{
Pascal Title
}
\seealso{
Naming of rasters in inputDir will be check with \code{\link{verifyFileStructure}}.
}

