% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerCreation.R
\name{layerCreation}
\alias{layerCreation}
\title{Creates all layers}
\usage{
layerCreation(masterstack, solradstack, var)
}
\arguments{
\item{masterstack}{rasterStack containing all precipitation, 
min temperature, max temperature and bioclimatic variables}

\item{solradstack}{rasterStack of monthly solar radiation}

\item{var}{vector of names of variables to generate, see Details.}
}
\value{
rasterStack
}
\description{
Generates all rasterLayers for one particular input dataset.
}
\details{
The function \code{\link{verifyFileStructure}} should be used to 
verify that the appropriate rasters are present in \code{masterstack}.

Possible variables to generate include:\cr

aridityIndexThornthwaite \cr
airidityIndexUNEP \cr
climaticMoistureIndex \cr
continentality \cr
embergerQ \cr
growingDegDays0 \cr
growingDegDays5 \cr
humidityIndex \cr
monthCountByTemp10 \cr
PETseasonality \cr
thermicityIndex \cr
minSummerPrecip \cr
maxSummerPrecip \cr
minWinterPrecip \cr
maxWinterPrecip \cr
minTempWarmest \cr
maxTempColdest \cr

If \code{var = 'all'}, then all of the variables will be generated.
}
\examples{
\dontrun{
# Find example rasters
rasterFiles <- list.files(system.file('extdata', package='envirem'), full.names=TRUE)

# create stack of temperature and precipitation rasters
# and stack of solar radiation rasters
solradFiles <- grep('solrad', rasterFiles, value=TRUE)
worldclim <- stack(setdiff(rasterFiles, solradFiles))
solar <- stack(solradFiles)

# generate all possible envirem variables
layerCreation(worldclim, solar, var='all')
}
}
\author{
Pascal Title
}
\seealso{
This function is called internally by \code{\link{generateRasters}}.
}

