% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namingScheme.R
\name{namingScheme}
\alias{namingScheme}
\title{List Naming Scheme}
\usage{
namingScheme()
}
\description{
Lists the naming scheme, either as default, or 
as modified by the user.
}
\details{
See \code{\link{assignNames}}.
}
\examples{
namingScheme()
}
