% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnames.R
\name{varnames}
\alias{varnames}
\title{List Naming Scheme}
\usage{
varnames()
}
\description{
Lists the naming scheme, either as default, or 
as modified by the user.
}
\details{
See \code{\link{assignNames}}.
}
\examples{
varnames()
}
