% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyFileStructure.R
\name{verifyFileStructure}
\alias{verifyFileStructure}
\title{Verify File Structure}
\usage{
verifyFileStructure(path, returnFileNames = TRUE, rasterExt = ".tif")
}
\arguments{
\item{path}{path to directory of rasters}

\item{returnFileNames}{logical, should file paths and names be returned}

\item{rasterExt}{file extension of rasters}
}
\value{
Prints messages to the console if problems are found. 
If \code{returnFileNames == TRUE}, then a vector of filenames is returned.
}
\description{
Ensures that the necessary files are present for other
functions to work properly.
}
\details{
This function searches for the following
in the directory specified by \code{path}:

19 bioclimatic variables (default naming: bio_)
	
12 precipitation rasters (default naming: precip_)

12 min temperature rasters (default naming: tmin_)

12 max temperature rasters (default naming: tmax_)

12 mean temperature rasters [optional] (default naming: tmean_)

12 solar radiation rasters (default naming: et_solrad_)

An equivalent function for checking the names of rasters in R can be 
found at \code{\link{verifyRasterNames}}. 

The naming of the variables is defined by default as shown above. 
Changes to the naming scheme can easily be made. See \code{link{?assignNames}}
for more information.

If mean temperature rasters are not detected, the raster creation functions
will create mean temperature by taking the mean of the min and max.
}
\examples{
# As there are no problems with these files, the list of files 
# will be returned.
verifyFileStructure(system.file('extdata', package='envirem'))
}
\author{
Pascal Title
}
