% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embergerQ.R
\name{embergerQ}
\alias{embergerQ}
\title{Emberger's pluviometric quotient}
\usage{
embergerQ(P, M, m, tempScale = 1, precipScale = 1)
}
\arguments{
\item{P}{rasterLayer, total annual precipitation}

\item{M}{rasterLayer, mean max temperature of the warmest month}

\item{m}{rasterLayer, mean min temperature of the coldest month}

\item{tempScale}{integer; scaling factor for the temperature data, see \link{envirem} for 
additional details.}

\item{precipScale}{integer; scaling factor for the precipitation data, see \link{envirem}
for additional details.}
}
\value{
rasterLayer in mm / degrees C
}
\description{
Calculate Emberger's pluviometric quotient.
}
\details{
\code{Q = 2000 P / [(M + m + 546.4) * (M - m)]}
}
\examples{
# Find example rasters
rasterFiles <- list.files(system.file('extdata', package='envirem'), full.names=TRUE)
env <- rast(rasterFiles)

embergerQ(env[['bio_12']], env[['bio_5']], env[['bio_6']], tempScale = 10)
}
\references{
Daget, P. (1977) Le bioclimat méditerranéen: analyse des formes climatiques 
par le système d’Emberger. \emph{Vegetatio}, \strong{34}, 87–103.
}
\author{
Pascal Title
}
